Controller code in one big file (Init.js)
